apply job controller
